Back, Front - Update Moods Views, Form, List, Card

+++ DailyMoodTracker/DailyMoodTracker/Controllers/MoodController.swift
@@ -80,6 +80,10 @@ class MoodController: ObservableObject {
             throw MoodError.deletionFailed("Impossible de supprimer une humeur par défaut.")
         }
         
+        guard !hasJournals(mood: mood) else {
+            throw MoodError.hasJournals("Impossible de supprimer une humeur utilisée dans un ou plusieurs journaux.")
+        }
+        
         context.delete(mood)
         do {
             try context.save()
@@ -177,6 +181,18 @@ class MoodController: ObservableObject {
         }
     }
     
+    // MARK: - hasJournals
+    // Vérifie si une humeur est utilisée dans un ou plusieurs journaux
+    func hasJournals(mood: Mood) -> Bool {
+        let journalController = JournalController(context: context)
+        do {
+            let journals = try journalController.getAll()
+            return journals.contains(where: { $0.mood == mood })
+        } catch {
+            return false
+        }
+    }
+    
     // MARK: - creatingADuplicate
     // Vérifie si une humeur créée un doublon (par défaut ou personnalisée) en ignorant un ID spécifique
     func creatingADuplicate(name: String, ignoredId: UUID) -> Bool {
@@ -206,6 +222,7 @@ enum MoodError: Error, LocalizedError {
     case saveFailed(String)
     case isDefaultMood(String)
     case alreadyExists(String)
+    case hasJournals(String)
     case creatingADuplicate(String)
 
     var errorDescription: String? {
@@ -221,6 +238,8 @@ enum MoodError: Error, LocalizedError {
         case .isDefaultMood(let message):
             return message
         case .alreadyExists(let message):
+            return message
+        case .hasJournals(let message):
             return message
         case .creatingADuplicate(let message):
             return message
